requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.3
django-redis==5.4.0

# Data Sources
pytrends==4.9.2
newsapi-python==0.2.7
requests==2.31.0

# AI/LLM
groq==0.9.0
openai==1.35.0

# Machine Learning
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5

# Utilities
django-htmx==1.17.3
whitenoise==6.7.0
Pillow==10.3.0
python-dateutil==2.9.0
